move the datalayer python type hints
